Sprint 5 done merging from dev to master

+++ HireTalent/DAO's/StudentDAO.swift
@@ -9,6 +9,8 @@
 import Foundation
 import Firebase
 import FirebaseAuth
+import FirebaseStorage
+import Kingfisher
 
 class StudentDAO{
     
@@ -33,6 +35,89 @@ class StudentDAO{
         }
     }
     
+    static func getImage(_ userId: String, imageView: UIImageView){
+           // Establish the connection with the database
+           let db = Firestore.firestore()
+           
+            
+          // Set a reference to the desired document
+          let empRef = db.collection("students").document(userId)
+          
+          empRef.getDocument { (document, error) in
+              
+              // If the specified document exist
+              if let document = document, document.exists {
+                  
+                  let empData = document.data()
+               guard let urlString = empData!["profilePicture"] as? String else{
+                   //print("error: does not have profile picture")
+                   return
+               }
+               guard let url = URL(string: urlString) else{
+                   print("error with url string ")
+                   return
+               }
+               let resource = ImageResource(downloadURL: url)
+               imageView.kf.setImage(with: resource, completionHandler: { (result) in
+                   switch result{
+                       
+                   case .success(_):
+                       print("Succesfully got image to the imageview")
+                   case .failure(_):
+                       print("ERROR coudn't get image to the imageview")
+                   }
+               })
+              } else {
+                  
+                  print("Did not find document for the url in getImage studentDAO")
+              }
+        }
+    }
+    
+    static func setImage(_ userId: String, _ image: UIImage, completion: @escaping ()-> Void){
+         print("Going to upload picture to the db")
+         let data = image.jpegData(compressionQuality: 0.1)
+         
+         
+         let imageName = UUID().uuidString
+         
+         let imageReference = Storage.storage().reference().child("studentImages").child(imageName)
+         
+         imageReference.putData(data!, metadata: nil){(metadata, error) in
+             
+             if error != nil{
+                 print("error puting image url")
+                 return
+             }
+             imageReference.downloadURL { (url, error) in
+                 if error != nil{
+                     print("error retrieving image url")
+                     return
+                 }
+                 if url == nil{
+                     print("error retrieving image url, it's null")
+                     return
+                 }
+                 let urlString = url?.absoluteString
+                 let db = Firestore.firestore()
+                 
+                     // Store the information in the database
+                     db.collection("students").document(userId).updateData([
+                         "profilePicture": urlString ?? "error"
+                     ]) { (error) in
+
+                         // Check for errors
+                         if error != nil {
+                             print("error uploading image to the db")
+                             return
+                         }
+                         print("succesfully added image to the db")
+                         completion()
+                         
+                    }
+            }
+        }
+    }
     
     // Insert the student data in the database.
     // It is used a callback because we depend of the 'result' provided by the setData() function.
@@ -104,11 +189,18 @@ class StudentDAO{
         // Establish the connection with the database
         let db = Firestore.firestore()
         
-        // Delete the student information in the database
-        db.collection("students").document("\(id)").delete()
-        
-        // Delete the information from authentication
-        Auth.auth().currentUser?.delete()
+        // Delete the student id from offers
+        JobOffersDAO.deleteInterestStudentId(id) { (error) in
+            if error != nil {
+                print("Error deleting student from all job offers")
+            } else {
+                // Delete the student information in the database
+                db.collection("students").document("\(id)").delete()
+                
+                // Delete the information from authentication
+                Auth.auth().currentUser?.delete()
+            }
+        }
     }
     
     // Get the user id
@@ -138,7 +230,6 @@ class StudentDAO{
         }
     }
        
-    
     // Get the general information of the student
     static func getStudent(_ userId: String, completion: @escaping(((String?), (Student?)) -> Void)) {
     
@@ -166,7 +257,7 @@ class StudentDAO{
                 student.semester = empData!["semester"] as? String ?? ""
                 student.experience = empData!["experience"] as? String ?? ""
                 student.notifications = empData!["notifications"] as? [String] ?? []
-                // Returns an object employer with all their data
+                // Returns an object student with all their data
                 completion(nil, student)
             } else {
                 
@@ -175,7 +266,6 @@ class StudentDAO{
             }
         }
     }
-    
     
     // Get a set of students given their student ids
     static func getStudents(_ studentIds: [String], completion: @escaping(([Student]?, [String]?) -> Void)) {
@@ -221,16 +311,16 @@ class StudentDAO{
                     // array to keep the same order in the students and in their ids
                     newStudentsIds.append(document.documentID)
                     
+                } else {
+                    db.collection("students").document(studentId).delete();
                 }
                 
                 if c == studentIds.count {
                     completion(students, newStudentsIds)
                 }
-                
             }
         }
     }
-    
     
     // Get a set of filtered students given their student ids and a speciality field
     static func getFilteredStudents(_ studentIds: [String], _ specialityField: String, completion: @escaping(([Student]?, [String]?) -> Void)) {
@@ -284,10 +374,28 @@ class StudentDAO{
                 numberOfStudents += 1
                 
                 if numberOfStudents == studentIds.count {
-                    print(students)
                     completion(students, newStudentsIds)
                 }
             }
+        }
+    }
+    
+    // Delete all the selected offer's notifications from a set of students caused by the deletion of an offer
+    static func deleteNotifications(_ offerKey: String, _ interestedStudents: [String]){
+        
+        // Add the offer key to an array because the arrayRemove() method receives an [Any]
+        var offerKeys: [Any] = []
+        offerKeys.append(offerKey)
+        
+        // Establish the connection with the database
+        let db = Firestore.firestore()
+        
+        for interestedStudent in interestedStudents {
+
+            // Delete the notifications for each student
+            db.collection("students").document(interestedStudent).updateData([
+                "notifications": FieldValue.arrayRemove(offerKeys)
+            ])
         }
     }
 }
